requirements.txt
# AI Co-Scientist Project Dependencies
# Phase 1: Foundation & Infrastructure

# Core LangChain packages for LLM integration
langchain-community==0.3.10
langchain-google-genai==2.0.8
langchain-core==0.3.28
langgraph==0.2.59

# Literature search (web search disabled to avoid rate limits)
requests==2.32.3

# Text processing and machine learning
sentence-transformers==3.3.1
scikit-learn==1.6.0
numpy==1.26.4

# Data handling and utilities
python-dotenv==1.0.1
dataclasses-json==0.6.7

# Standard library replacements and utilities
typing-extensions==4.12.2
urllib3==2.2.3

# Development and testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Optional: Jupyter notebook support for development
jupyter==1.1.1
ipython==8.30.0

# XML parsing for literature search
lxml==5.3.0

# Async support
aiohttp==3.11.11